address clippy lints

+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod particle;
 pub mod vector;
 
-pub type Real = f32;
+pub type Real = f64;
 
 pub use self::{particle::*, vector::*};
